Enhanced calculator UI/UX, added advanced functions, and integrated a login form.
