--- ContactForm.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>    
-<head>    
-<title>Contact Form</title>    
-<style>
-.contact{  
-        width: 382px;  
-        overflow: hidden;  
-        margin: auto;  
-        margin: 20 0 0 450px;  
-        padding: 85px;  
-        background: orange;  
-        border-radius: 21px ;  
-}  
-h2{  
-    text-align: center;  
-    color: lightgreen;  
-    font-size:45px;  
-}  
-label{  
-    color: blue;  
-    font-size: 21px;  
-}  
-.Name{  
-    width: 300px;  
-    height: 30px;  
-    border: none;  
-    border-radius: 5px;  
-    padding-left: 9px;  
-}
-.button{
-  padding: 5px 13px;
-  display: inline-block;
-  text-align: center;
-  font-size: 25px;
-  cursor: pointer;
-  font-family: Algerian;
-}
-.button1 {
-  background-color: red;
-  color: lightgreen;
-}
-.button1:hover {
-  background-color: green;
-  color: white;
-}
-
-</style>    
-</head>    
-<body background="P3.jpg">    
-<h2><u>Contact Us</u></h2><br>    
-<div class="login">    
-<form class="contact" method="get" action="contact.php">    
-<label><b>Your Name     
-</b>    
-</label>    
-<input type="text" name="Yname" class="Name">    
-<br><br>    
-<label><b>Your Mail     
-</b>    
-</label>    
-<input type="text" name="Mail" class="Name"> 
-<br><br>
-<label><b>Your Phone No.  
-</b>    
-</label>   
-<input type="text" name="Phoneno." class="Name">
-<br><br>
-<label><b>Your Message     
-</b>    
-</label>    
-<input type="text" name="Message" class="Name">   
-<br><br><br>
-&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp<input type="submit" class="button button1" border="1" value="Submit"/>
-</form>     
-</div>    
-</body>    
-</html>
