+++ README.md
@@ -1,3 +1,69 @@
-# web-page
+# Web-Page Repository
 
-<h8>Star and Forking this repo is the must to get accept you pull requests</h8>
+Welcome to the `web-page` repository! This repository is a collection of diverse code snippets, games, algorithms, and web-related files. If you are interested in contributing or simply want to explore, you've come to the right place!
+
+## Table of Contents
+
+- [Features](#features)
+- [How to Contribute](#how-to-contribute)
+- [Important Notes](#important-notes)
+- [License](#license)
+
+## Features
+
+Here's what you can find in this repository:
+
+1. **Games**:
+   - Snake Game
+   - PING PONG
+   - (and many more in the future!)
+
+2. **Algorithms & Programs**:
+   - Greedy Knapsack (C)
+   - Guess The Number (Java)
+   - Longest Substring Without Repeating Characters (Python)
+   - Binary Search (Python)
+   - Bubble Sort (Python & C++)
+   - Queue Using Array (C++)
+   - LCS (C)
+   - SCS (C)
+   - (and others)
+
+3. **Web Development**:
+   - Contact Form (HTML)
+   - Scientific Calculator (HTML)
+   
+4. **Miscellaneous Scripts & Files**:
+   - IRON MAN (Python)
+   - REVERSE INTEGER (Python)
+   - SPELL CORRECTOR (Python)
+   - (and a few more!)
+
+## How to Contribute
+
+If you're interested in contributing to this repository, please follow the steps below:
+
+1. **Star & Fork**: Before making any contribution, make sure to star and fork this repository. This is a must!
+2. **Clone**: Clone your forked copy to your local system.
+3. **Branch**: Create a new branch. Use a clear name for your branch, for instance: `add-xyz-algorithm`.
+4. **Commit**: Make necessary changes and commit those changes.
+5. **Push**: Push changes to GitHub in your forked repository.
+6. **Pull Request**: Create a pull request (from your forked repo to the main repo) and patiently wait for it to be reviewed.
+
+## Important Notes
+
+- Please ensure your code is clean, well-documented, and error-free.
+- Update the README if you add new content.
+- If you spot a bug or have an improvement suggestion, raise an issue before making a pull request.
+  
+## License
+
+This repository is open-sourced under the MIT License. For more details, see [LICENSE](LICENSE).
+
+---
+
+**Happy Coding!** 🚀
+
+--- 
+
+**Acknowledgment**: Special thanks to @itsdchanda and all other contributors for making this repository vibrant and evolving!
