--- SciCalc.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Scientific Calculator</title>
-    <style>
-        body {
-            font-family: 'Arial', sans-serif;
-            text-align: center;
-            margin: 20px;
-        }
-
-        #calculator {
-            width: 300px;
-            margin: 0 auto;
-            padding: 10px;
-            border: 1px solid #ccc;
-            border-radius: 5px;
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-        }
-
-        input {
-            width: 100%;
-            margin-bottom: 10px;
-            padding: 8px;
-            box-sizing: border-box;
-        }
-
-        button {
-            width: 48px;
-            height: 48px;
-            font-size: 16px;
-            margin: 5px;
-            cursor: pointer;
-        }
-    </style>
-</head>
-<body>
-
-<div id="calculator">
-    <input type="text" id="display" readonly>
-
-    <button onclick="appendToDisplay('1')">1</button>
-    <button onclick="appendToDisplay('2')">2</button>
-    <button onclick="appendToDisplay('3')">3</button>
-    <button onclick="appendToDisplay('+')">+</button>
-    <br>
-
-    <button onclick="appendToDisplay('4')">4</button>
-    <button onclick="appendToDisplay('5')">5</button>
-    <button onclick="appendToDisplay('6')">6</button>
-    <button onclick="appendToDisplay('-')">-</button>
-    <br>
-
-    <button onclick="appendToDisplay('7')">7</button>
-    <button onclick="appendToDisplay('8')">8</button>
-    <button onclick="appendToDisplay('9')">9</button>
-    <button onclick="appendToDisplay('*')">*</button>
-    <br>
-
-    <button onclick="appendToDisplay('0')">0</button>
-    <button onclick="appendToDisplay('.')">.</button>
-    <button onclick="clearDisplay()">C</button>
-    <button onclick="appendToDisplay('/')">/</button>
-    <br>
-
-    <button onclick="calculate()">=</button>
-    <button onclick="appendToDisplay('Math.sqrt(')">√</button>
-    <button onclick="appendToDisplay('Math.sin(')">sin</button>
-    <button onclick="appendToDisplay('Math.cos(')">cos</button>
-    <button onclick="appendToDisplay('Math.tan(')">tan</button>
-</div>
-
-<script>
-    function appendToDisplay(value) {
-        document.getElementById('display').value += value;
-    }
-
-    function clearDisplay() {
-        document.getElementById('display').value = '';
-    }
-
-    function calculate() {
-        try {
-            document.getElementById('display').value = eval(document.getElementById('display').value);
-        } catch (error) {
-            document.getElementById('display').value = 'Error';
-        }
-    }
-</script>
-
-</body>
-</html>
