+++ Web Development/styles.css
@@ -1,0 +1,75 @@
+/* Universal styling */
+body {
+    font-family: 'Roboto', sans-serif;
+    background-color: rgba(0,0,0,0.6);
+    background-image: linear-gradient(45deg, #FF4466, #4466FF);
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center center;
+}
+
+h2 {
+    text-align: center;
+    color: white;
+    font-size: 45px;
+    margin-bottom: 30px;
+    text-decoration: bold;
+}
+
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    min-height: 100vh;
+}
+
+/* Contact form styling */
+.contact {
+    width: 400px;
+    padding: 40px;
+    background-color: rgba(255, 255, 255, 0.1);
+    backdrop-filter: blur(5px);
+    border: 2px solid rgba(255, 255, 255, 0.2);
+    border-radius: 21px;
+    box-shadow: 10px 10px 30px rgba(0, 0, 0, 0.2);
+}
+
+label {
+    display: block;
+    color: #FFF;
+    font-size: 21px;
+    margin-bottom: 10px;
+}
+
+.Name {
+    width: 100%;
+    height: 30px;
+    padding: 5px 10px;
+    border: 2px solid rgba(255, 255, 255, 0.7);
+    border-radius: 5px;
+    margin-bottom: 20px;
+    background-color: rgba(0,0,0,0.3);
+    color: #FFF;
+}
+
+.button {
+    width: 100%;
+    padding: 10px;
+    font-size: 20px;
+    cursor: pointer;
+    transition: background-color 0.3s, color 0.3s;
+    border: none;
+    border-radius: 5px;
+    box-shadow: 5px 5px 15px rgba(0, 0, 0, 0.2);
+}
+
+.button1 {
+    background-color: rgba(255, 255, 255, 0.7);
+    color: rgb(195, 40, 130);
+}
+
+.button1:hover {
+    background-color: rgba(255, 158, 231, 0.7);
+    color: white;
+}
